[MS-1519] Consolidate the two isValidGuid() functions into one
