+++ feature/client-api/src/main/java/com/simprints/feature/clientapi/ClientApiViewModel.kt
@@ -8,9 +8,9 @@ import androidx.lifecycle.viewModelScope
 import com.simprints.core.livedata.LiveDataEvent
 import com.simprints.core.livedata.LiveDataEventWithContent
 import com.simprints.core.livedata.send
-import com.simprints.core.tools.extentions.isValidGuid
 import com.simprints.core.tools.extentions.toJsonElementMap
 import com.simprints.core.tools.time.TimeHelper
+import com.simprints.core.tools.utils.isValidGuid
 import com.simprints.feature.clientapi.exceptions.InvalidRequestException
 import com.simprints.feature.clientapi.extensions.toMap
 import com.simprints.feature.clientapi.mappers.request.IntentToActionMapper
