--- infra/core/src/main/java/com/simprints/core/tools/extentions/String.ext.kt
@@ -1,10 +1,0 @@
-package com.simprints.core.tools.extentions
-
-import java.util.UUID
-
-fun String.isValidGuid() = try {
-    UUID.fromString(this)
-    true
-} catch (_: IllegalArgumentException) {
-    false
-}
